Refactored Aurn into extract and process
