+++ environmental_data_modules/__init__.py
@@ -1,6 +1,8 @@
 from .environment_module import EnvironmentModule
 from .met_extractor import MetExtractor
+from .aurn_module import AurnModule
 from .aurn_extractor import AurnExtractor
 from .post_processor import PostProcessor
 from .met_post_processor import MetPostProcessor
 from .aurn_post_processor import AurnPostProcessor
+
